Lesson 4 - class file 2.9.1.html: List elements
Lesson on the different types of list elements. <ol>(ordered List) element is used to created a descending numbered list. <ul>(unordered list) element uses bullet points to list content. <li>(list item) element is wrap in side or either <ul> or <ol> and wrap the content that is being displayed as a list.