Lesson 6 - class file 2.11.html: Forms
You use the <form> element to house the structure of a form. The Form element has two attributes, action and method.

The action attributes specifies where the form will be submitted to for processing.
The method attribute specifies how the form will be submited. You can either use the get or post method with is more desirable and safe.

the label element helps with creating label for the text boxes of the inputs.
input element creates the input box and has for attributes.

the type attributes helps with deciding on what kind of input to expect.
the name attribute
id attributes allow the for attribute on the label to link to the input box when the text it displays is clicked.
autofocus attribute.